requirements.txt
python-dotenv>=1.0.0
python-dateutil>=2.8.2
supabase>=2.3.1
langchain-groq>=0.1.0
langchain-core>=0.1.0
langchain>=0.1.0